% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.hs.genes.R
\docType{data}
\name{biomart.hs.genes}
\alias{biomart.hs.genes}
\alias{hs.genes}
\title{Biomart Homo Sapiens gene annotation data (GRCh38.p3)}
\usage{
data(biomart.hs.genes)
}
\description{
Biomart homo sapiens gene annotation data downloaded from
biomart website (http://www.ensembl.org/biomart/, 2015-10-09). Can be used for a fast annotation of genes
compared to R/Bioconductor biomaRt package.
}
\examples{
\donttest{

}
}
