% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_Results-class.R
\docType{class}
\name{DE_Results-class}
\alias{DE_Results-class}
\alias{DE_Results}
\alias{$<-,DE_Results-method}
\title{DE_Results Class}
\usage{
DE_Results(x)

\S4method{$}{DE_Results}(x, name) <- value
}
\arguments{
\item{x}{A data frame or any object coercible to a \code{DFrame}.}
}
\value{
A \code{DE_Results} object.
}
\description{
This class extends the \code{\link[S4Vectors]{DFrame}} class from the
S4Vectors package, designed specifically to hold results from
differential expression analysis, ensuring specialized methods can be
applied.

Creates an instance of the \code{DE_Results} class. This function
validates the input and ensures proper construction of the S4 object.
}
