% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DE_Results-class}
\alias{DE_Results-class}
\alias{DE_Results}
\title{DE_Results object and constructor}
\usage{
DE_Results(DataFrame)
}
\arguments{
\item{DataFrame}{a DataFrame of lm_results.}
}
\value{
a DE_Results object
}
\description{
DE_Results extends the DataFrame class of S4Vectors. It is used by \code{\link{lm_results}}
to wrap up the results table. This constructor function would not typically be used by "end users".
}
