% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressed.R
\name{expressed}
\alias{expressed}
\alias{summary.expressed}
\title{Expressed genes in each cell population}
\usage{
expressed(x, groups, cutoff = 1)

\method{summary}{expressed}(object, ...)
}
\arguments{
\item{x}{matrix or numeric data frame containing gene expression data. Rows are genes and columns are samples.}

\item{groups}{a factor containing the group assignement of samples}

\item{cutoff}{a numeric value specifying the threshold above which a gene is declared as expressed.}

\item{object}{an object of class expressed}

\item{...}{additional arguments.}
}
\value{
\itemize{
\item expressed() returns an object of class "expressed". Columns are samples or groups and rows are genes. Values can be 1 (if expressed)
or 0 (if not expressed).
\item summarize() returns a list containing the following components:
 \itemize{
   \item exprs_in_all: The number of expressed genes in all groups
   \item exprs_in_one: The number of expressed genes in at least one group
   \item exprs_in_each: a vector containing the number of genes expressed in each group
   \item exprs_in_specific: the number of genes expressed only in specific group
 }
}
}
\description{
Identification of expressed genes in each cell population.
Expressed genes in a given cell population is defined as genes with average expression above a given cutoff.
}
