% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regulators.R
\name{get_human_regulators}
\alias{get_human_regulators}
\title{Identification of a transcription regulators in a list of genes}
\usage{
get_human_regulators(ids, id_type = "ensembl", keep.all = TRUE)
}
\arguments{
\item{ids}{a vector containing gene Ensembl Ids ("NSG00000130816") or gene names ("DNMT1").
if ids is a matrix or a data.frame, then the row names are used as ids.}

\item{id_type}{the type of the ids. Possible values are "ensembl" or "name".}

\item{keep.all}{a logical value. If TRUE, all the genes are kept in the final result. If FALSE, only
transcription factors/epigenetic enzymes are returned}
}
\value{
Returns an object of class data.frame containing the following columns:
\itemize{
\item id: gene ids
\item TF: with a value of 1 (the genes is transcription factor) or 0 (the gene is not a transcription factor).
\item EE_class: specifies the epigenetic enzyme class of the genes.
}
}
\description{
Identification of DNA-binding transcription factors and/or epigenetic enzymes in a list of genes.
Only data for human are available (see ?human.tf and ?human.ee).
}
