% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDiff_Results-class.R
\docType{class}
\name{GetDiff_Results-class}
\alias{GetDiff_Results-class}
\alias{GetDiff_Results}
\title{GetDiff_Results Class}
\usage{
GetDiff_Results(x)
}
\arguments{
\item{x}{A data frame or any object coercible to a \code{DE_Results} object.}
}
\value{
A valid \code{GetDiff_Results} object.
}
\description{
This class is a specialized subclass of \code{\link{DE_Results}},
intended for results that have undergone a specific "get_diff" transformation
or analysis step. It inherits all slots and methods from \code{DE_Results}.

Creates an instance of the \code{GetDiff_Results} class. It first coerces the
input to a \code{DE_Results} object and then re-casts it as the specialized
subclass.
}
