% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.ee.R
\docType{data}
\name{human.ee}
\alias{human.ee}
\alias{h.ee}
\title{Human epigenetic enzymes}
\format{A data frame with 4 columns and 212 rows coorresponding to genes.}
\source{
Yang Z, Jones A, Widschwendter M, Teschendorff AE.
An integrative pan-cancer-wide analysis of epigenetic enzymes reveals universal patterns of epigenomic deregulation in cancer.
Genome Biol. 2015 Jul 14;16:140. doi: 10.1186/s13059-015-0699-9.
}
\usage{
data(human.ee)
}
\description{
A list of 212 epigenetic enzymes, including writers, readers, editors
and erasers of the most important epigenetic marks.
}
\examples{
\donttest{
 data(human.ee)
 head(human.ee)
}
}
