% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.tf.R
\docType{data}
\name{human.tf}
\alias{human.tf}
\alias{h.tf}
\title{Census of human transcription factors}
\format{A data frame with 6 columns and 1391 rows coorresponding to genes.}
\source{
Juan M. Vaquerizas, Sarah K. Kummerfeld, Sarah A. Teichmann & Nicholas M. Luscombe.
A census of human transcription factors: function, expression and evolution.
Nature Reviews Genetics 10, 252-263 (April 2009) | doi:10.1038/nrg2538
}
\usage{
data(human.tf)
}
\description{
Census list of 1391 human DNA-binding transcription factors as described by Vaquerizas et al., 2009.
}
\examples{
\donttest{
 data(human.tf)
 head(human.tf)
}
}
