% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{ggbarplot}
\alias{ggboxplot}
\alias{gglineplot_by}
\title{Graphs}
\usage{
ggbarplot(
  x,
  title = "",
  xlab = "",
  ylab = "",
  fill = "steelblue",
  color = "steelblue",
  sort.value = c("none", "desc", "asc"),
  top = Inf,
  ...
)

ggboxplot(data, xName, yName, fill = "steelblue", color = "black")

gglineplot_by(
  exprs,
  groups,
  log_it = TRUE,
  title = "",
  labels = FALSE,
  xlab = "Groups",
  ylab = "log2 fold change compared to the control",
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{title, xlab, ylab}{labels for the graph}

\item{fill}{a fill color}

\item{color}{for line or point}

\item{sort.value}{a string specifying whether x should be sorted or not.
Allowed values are "none" (no sorting), "asc" (for ascending) or "desc" (for descending)}

\item{top}{a numeric value specifing the top elements to be shown}

\item{...}{additional arguments}

\item{data}{a data frame}

\item{xName, yName}{variables to be used for x and y, respectively}

\item{exprs}{gene expression data (rows are genes and columns are samples)}

\item{groups}{a vector containing the group of each samples}

\item{log_it}{a logical value. If TRUE the data are log2 transformed.}

\item{labels}{logical value. If TRUE point labels are shown.}
}
\description{
Create graphs
}
\examples{
\donttest{
data(mtcars)
ggbarplot(mtcars[, "mpg"], top = 10)
}
}
