% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MA.R
\name{plot_MA}
\alias{plot_MA}
\alias{plot.DESeqResults}
\alias{plot.DE_Results}
\title{MA-plot from means and log fold changes}
\usage{
plot_MA(
  x,
  fdr = 0.05,
  fc = 1.5,
  detection_call = NULL,
  ylim = NULL,
  colNonSig = "darkgray",
  colUp = "#BC002A",
  colDown = "#16614E",
  colLine = "black",
  log = "x",
  cex = 0.45,
  xlab = "Mean expression",
  ylab = "Log2 fold change",
  addlegend = TRUE,
  highlight = 10,
  ...
)

\method{plot}{DESeqResults}(x, ...)

\method{plot}{DE_Results}(x, ...)
}
\arguments{
\item{x}{an object of class DESeqResults, get_diff, DE_Results, matrix or data frame containing the columns
baseMean, log2FoldChange, and padj. Rows are genes.
\itemize{
\item baseMean: the mean expression of genes in the two groups.
\item log2FoldChange: the log2 fold changes of group 2 compared to group 1
\item padj: the adjusted p-value of the used statiscal test.
}}

\item{fdr}{Accepted false discovery rate for considering genes as differentially expressed.}

\item{fc}{the fold change threshold. Only genes with a fold change >= fc and padj <= fdr are
considered as significantly differentially expressed.}

\item{detection_call}{a numeric vector with length = nrow(x),
specifying if the genes is expressed or not. Default is NULL. If detection_call column is available in x,
it will be used.}

\item{ylim}{The limits for the y-axis.
If missing, an attempt is made to choose a sensible value.
Dots exceeding the limits will be displayed as triangles at the limits, pointing outwards.}

\item{colNonSig, colUp, colDown}{colors to be used for non-significant,
up-regulated and down-regulated data points, respectively.}

\item{colLine}{colour to use for the horizontal (y=0) line.}

\item{log}{which axis/axes should be logarithmic; will be passed to plot.}

\item{cex}{point size}

\item{xlab, ylab}{x and y axis labels}

\item{addlegend}{a logical value. If TRUE a legend is added to the plot.}

\item{highlight}{highlight top genes. Default value is 10 for top 10 genes.}

\item{...}{other parameters to be passed to the function plot}
}
\value{
returns a base plot
}
\description{
Make MA-plot which is a scatter plot of log2 fold changes (on the y-axis)
versus the mean expression signal (on the x-axis).
}
