% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepProfiler.R
\name{stepProfiler}
\alias{stepProfiler}
\alias{pairewise_compare}
\alias{one_step_up}
\alias{one_step_down}
\alias{impulsed_up}
\alias{impulsed_down}
\alias{plot.stepgenes}
\alias{summary.stepgenes}
\alias{pathways}
\title{StepProfiler: Extract temporally Up- and Down- regulated genes using DESeq2.}
\usage{
pairewise_compare(dds, factor, grps, alpha = 0.05, fc = 1,
  active_exprs = 1, count.norm = NULL, exprs.norm = NULL,
  verbose = TRUE, ...)

one_step_up(pwc, fc = 1.5, verbose = TRUE)

one_step_down(pwc, fc = 1.5, verbose = TRUE)

impulsed_up(pwc, fc = 1.5, verbose = TRUE)

impulsed_down(pwc, fc = 1.5, verbose = TRUE)

\method{plot}{stepgenes}(x, transformby = c("firststep", "basemean",
  "standardize", "none"), size = c(0.2, 1), getRegFunc = NULL,
  print_plot = TRUE, ...)

\method{summary}{stepgenes}(object, ...)

pathways(object, gn_id_type = "ENSEMBL", pvalueCutoff = 0.05,
  OrgDb = "org.Hs.eg.db", verbose = TRUE)
}
\arguments{
\item{dds}{an object of class DESeqDataSet containing the diff_exprs of the function DESeq().}

\item{factor}{the name of a factor in the design formula (e.g factor = "group")}

\item{grps}{a vector containing at list 2 groups to be compared.}

\item{alpha}{false discovery rate}

\item{fc}{positive value specifying fold change threshold. Must be at least 1.
Genes with fc > 1 are kept.}

\item{active_exprs}{a numeric value specifying detected expression cutoff.
If not NULL, any gene with an average expression above this cutoff, in at least one group,
is declared as actively expressed. Used for defining presence and absence detection call.
If not NULL, a column named detection_call is added to the result. Possible values for detection_call
is 0 (the gene is not actively expressed) or 1 (the gene is actively expressed).}

\item{count.norm}{a data frame containing the normalized count as provided by DESeq2.
Only used for computing the detection call when active_expression is not NULL.}

\item{exprs.norm}{a data frame containing the normalized gene expression data. An alias of count.norm.}

\item{verbose}{logical value; if TRUE the progression is shown.}

\item{...}{other arguments}

\item{pwc}{an object of class "pairewise_compare" as returned by the function pairewise_compare().}

\item{x}{an object of class step_genes to be plotted}

\item{transformby}{a character vector specifying how to transform the expression signal before plotting.
Possible values are:
\itemize{
\item standardize: the rlog data is standardized, before plotting, using the function scale().
\item firstep: the log2 fold change expression compared to the first group is displayed
\item none: the mean rlog expression is displayed
}}

\item{size}{a numeric vector of length 2 containing the size of lines and points, respectively.}

\item{getRegFunc}{a function for identifying regulators (transcription factor/epigenetic enzymes) in a list of genes.
A possible value is get_human_regulators.}

\item{print_plot}{a logical value used in the function plot.stepgenes().
If TRUE ggplots are printed. If FALSE, a list, which elements are ggplots, is returned instead of being printed.}

\item{object}{an object of class step_genes to be plotted}

\item{gn_id_type}{the type of gene ids. Allowed values are "ENSEMBL"
"ENTREZID", "SYMBOL", "UNIGENE", "GENENAME" or "UCSCKG".}

\item{pvalueCutoff}{Cutoff value of pvalue}

\item{OrgDb}{annotation database. Default is "org.Hs.eg.db" for human.}
}
\value{
\bold{pairewise_compare()} returns a list containing the following elements:
\itemize{

\item significance: a data frame containing the pairewise comparison results.
Each column corresponds to the significance of a given comparison. Rows are genes wich can have -1, 0 or 1 values for
each column. -1 = significantly downregulated; 0 = not significant; 1 = significantly upregulated.
A significant deregulated gene is a gene with active expression AND padj < = alpha AND absolute fold change >= fc.
\item foldchanges: a data frame containing log2 fold changes from the different comparisons.
\item padj: a data frame containing the adjusted p-values from the different comparisons.
\item groups:  the differents groups used for the pairewise comparisons.
This corresponds to levels(samples$group).
\item steps: transition steps from group_i to group_i+1
\item nstep: the number of steps
\item steps_mean_fc: a numeric vector containing the mean of the absolute fold change during transitions.
Can be useful for sorting genes.
\item dds: the dds used during pairewise comparison.
\item count.norm: the normalized count as provided by DESeq2
\item rlog.data: 'regularized log' transformed data as provided by DESeq2
\item samples: samples data as provided by colData(dds)
\item refs: a numeric vector containing the mean normalized expression of each gene in the first group.
This is used when plotting the results with argument transformby = "firstgroup" in order to display the
log2 fold change expression in the other groups compared to the first group as reference samples.
\item detection_call a vector containing 1 or 0 specifying whether the gene is actively expressed or not in at
least one group.
This computed according to the value of the argument active_exprs.
\item steps_mean_fc the mean of the absolute fold change during transitions. Can be useful for sorting genes.
\item baseMean the mean normalized count for all samples.
}
\bold{one_step_*(), impulsed_*()} returns an object of class stepgenes which is a list of sub-lists.
Each sub-list corresponds to a given step and contains the following elements:
\itemize{
\item exprs: a data frame containing the normalized gene expression values.
Columns are samples and rows are genes.
\item foldchanges, padj, refs, baseMean: see the description above.
\item group: a factor containing the group assignment of samples
}
Genes are ordered according to the mean fold change expression during all transitions,
so that the most relevant genes are shown at the top the data table.

\bold{plot()}: Plots step genes using ggplot2.\cr\cr

\bold{summary()}: Returns a data frame which columns are:
\itemize{
\item id: gene IDs
\item status: the expression status of a gene at a given step
\item step: the step at which the expression status changes
\item group: Numeric coding of the step column
\item max_fcs and min_fcs: the maximum and the minimum fold change expression at each transition, respectively.
}
}
\description{
stepProfiler can be used to identify four temporal gene expression patterns that we refer to as
one-step-up, one-step-down, impulsed-up, impulsed-down.
\itemize{
\item pairewise_compare(): paire-wise comparison between all groups.
\item one_step_up(): Low before a step i and increase gradually, at step i+1, during all the next transition steps.
\item one_step_down(): High before a step i and  decrease gradually, from step i+1,  during all the next transition steps.
\item one_step_up_sustained(): Low before a step i, upregulated, from step i+1, and remain stable during all the next transition steps:
the maximum fold change between the step i+1 and the following steps must be in [-1.4, 1.4] (stability fold change).
\item one_step_down_sustained(): Highly expressed before step i, downregulated at step i+1 and remain stable
during all the next transition steps: the maximum fold change between the current step and the next step must be in [-1.4, 1.4].
\item impulses_up(): upregulated gradually and significantly during all the previous transitions (i--) until the current step i,
then decrease significantly and gradually during all the next transition steps. There is an impulse signal at
step i with a minimum fc = 1.5. This expression profile is also called as two-step-down-up.
\item impulse_up_sustained(): Low and stable during the previous and the next transitions,
significant impulse at step i with a minimum fc = 1.5.
\item impulse_down(): downregulated significantly and gradually during all the previous transitions (i--)
until the current step i, then increase significantly and gradually during all the next transition steps.
There is an impulse signal at step i with a minimum fc = 1.5. This expression profile is also called as two-step-down-up.
\item impulse_down_sustained(): High and stable during the previous and the next transitions,
significant impulse at step i with a minimum fc = 1.5.
\item plot(): Draw step genes using ggplot2
}
}
