% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarizeby}
\alias{summarizebys}
\title{Summarize gene expression data}
\usage{
summarize(x)

summarizeby(x, groupby, fun, melt = FALSE)

summarizebys(x, groupby, stat = c("mean", "sd"))
}
\arguments{
\item{x}{matrix or numeric data frame containing gene expression data. Rows are genes and columns are samples.}

\item{groupby}{a factor variable specifying sample groups}

\item{fun}{a statistical function such as mean, median etc}

\item{melt}{logical value. If TRUE, then the data is summarized by groups
and melted using reshape2 package}

\item{stat}{a vector containing the name of the statistic summaries you want to compute.
Possible values are the combination of the following statistics:
c("mean", "sd", "min", "max", "median", "mad", "iqr")}
}
\value{
\itemize{
\item summarize() returns a data frame containing some descriptive statistics for each genes including:
the minimum (Min), the first quartile (Q1), the median (Med), the mean (Mean), the third quartile (Q3),
the maximum (Max), the standard deviation (SD), the interquartile range (IQR), the median absolute deviation (MAD).
\item summarizeby() retuns a data frame containing the summarized expression data by groups.
}
}
\description{
Summarize gene expression data by computing some descriptive statistics.
\itemize{
\item summarize(): computes descriptive statistics for expression data.
\item summarizeby(): summarizes expression data by groups and by \bold{one descriptive statistic}.
For example computes, the
average expression of each gene by sample groups.
\item summarizebys(): summarizes expression data by groups
and by \bold{multiple descriptive statistics} at the same time. For example computes, the
mean expression and the SD of each gene by sample groups.
}
}
