% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezrnaseq.R
\name{import_rnaseq}
\alias{import_rnaseq}
\alias{compare_two_groups}
\alias{export_significant}
\title{Import RNAseq Data}
\usage{
import_rnaseq(directory = "data")

compare_two_groups(data, grp1, grp2, fold_change = 2)

export_significant(results)
}
\arguments{
\item{directory}{path to the directory containing the RNAseq data. This directory
should contain the following files:
\itemize{
\item{count.txt}: containsraw count data
\item{samples.txt}: contains sample groups
\item {annotation.txt}: contains the gene annotation file.
}}

\item{data}{a list containing gene expression data plus samples and gene annotation.}

\item{grp1}{group 1 used as numerator for calculating the fold change.}

\item{grp2}{group 2 used as denominator for calculating the fold change.}

\item{fold_change}{fold change to be considered}

\item{results}{results returned by compare_groups}
}
\description{
Imports RNAseq data contained in a directory.
}
\section{Functions}{
\itemize{
\item \code{import_rnaseq}: Import RNAseq data

\item \code{compare_two_groups}: Compare two groups

\item \code{export_significant}: Export significant results
}}

